+++ packages/core/android/src/main/java/io/sentry/react/RNSentryReactFragmentLifecycleTracer.java
@@ -16,6 +16,8 @@ import io.sentry.ILogger;
 import io.sentry.SentryLevel;
 import io.sentry.android.core.BuildInfoProvider;
 import io.sentry.android.core.internal.util.FirstDrawDoneListener;
+import java.util.HashMap;
+import java.util.Map;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
@@ -24,6 +26,7 @@ public class RNSentryReactFragmentLifecycleTracer extends FragmentLifecycleCallb
   private @NotNull final BuildInfoProvider buildInfoProvider;
   private @NotNull final Runnable emitNewFrameEvent;
   private @NotNull final ILogger logger;
+  private final Map<Fragment, EventDispatcherListenerWrapper> listenerWrapperMap = new HashMap<>();
 
   public RNSentryReactFragmentLifecycleTracer(
       @NotNull BuildInfoProvider buildInfoProvider,
@@ -83,18 +86,40 @@ public class RNSentryReactFragmentLifecycleTracer extends FragmentLifecycleCallb
       return;
     }
 
-    final @NotNull Runnable emitNewFrameEvent = this.emitNewFrameEvent;
-    eventDispatcher.addListener(
-        new EventDispatcherListener() {
+    EventDispatcherListenerWrapper listenerWrapper =
+        new EventDispatcherListenerWrapper(eventDispatcher) {
           @Override
           public void onEventDispatch(Event event) {
             if ("com.swmansion.rnscreens.events.ScreenAppearEvent"
                 .equals(event.getClass().getCanonicalName())) {
-              eventDispatcher.removeListener(this);
+              this.dispatcher.removeListener(this);
+              listenerWrapperMap.remove(f);
               FirstDrawDoneListener.registerForNextDraw(v, emitNewFrameEvent, buildInfoProvider);
             }
           }
-        });
+        };
+
+    eventDispatcher.addListener(listenerWrapper);
+    listenerWrapperMap.put(f, listenerWrapper);
+  }
+
+  @Override
+  public void onFragmentViewDestroyed(@NonNull FragmentManager fm, @NonNull Fragment f) {
+    super.onFragmentViewDestroyed(fm, f);
+    EventDispatcherListenerWrapper listenerWrapper = listenerWrapperMap.get(f);
+    if (listenerWrapper != null && listenerWrapper.dispatcher != null) {
+      listenerWrapper.dispatcher.removeListener(listenerWrapper);
+      listenerWrapperMap.remove(f);
+    }
+  }
+
+  abstract static class EventDispatcherListenerWrapper implements EventDispatcherListener {
+
+    protected final EventDispatcher dispatcher;
+
+    public EventDispatcherListenerWrapper(EventDispatcher dispatcher) {
+      this.dispatcher = dispatcher;
+    }
   }
 
   private static @Nullable EventDispatcher getEventDispatcherForReactTag(